+++ releasenotes.md
@@ -1,5 +1,6 @@
 Whats New in v11.0
 --------------------------
+- Making ingredient profile customizable
 - Added maximum allowed usage for a formula
 - Added a button to view ingredient data in PubChem
 - Changed the way the ingredient is handled when exists in a formula but not in the database. Instead of auto creating it when clicked, it presents options to create it, search in PV Online or import via JSON